win32: fix 2 compiler warning nits in win32 port
Reported-by: Gisle Vanem <gisle.vanem@gmail.com>

+++ contrib/ports/win32/include/arch/cc.h
@@ -103,7 +103,7 @@ void lwip_win32_platform_diag(const char *format, ...);
 #endif
 
 extern unsigned int lwip_port_rand(void);
-#define LWIP_RAND() (lwip_port_rand())
+#define LWIP_RAND() ((uint32_t)lwip_port_rand())
 
 #define PPP_INCLUDE_SETTINGS_HEADER
